Wastewater management and household infrastructure in Brazil - Demographic Census 2010

This dataset about wastewater management and household infrastructure from various Brazilian regions provides insights into wastewater disposal habits, water sources, bathroom facilities, and sanitation infrastructure.

Usage

wsabrazil

Format

A tibble with 1965 rows and 19 variables

sector_code

sector code

MR_name

metropolitan region name

municipality_code

municipality code

municipality_name

municipality name

sector_situation

location type: urban or rural

sector_type

living conditions: 1 stands for poor housing conditions

avg_income

average nominal monthly income of permanent private households

total_households

number of permanent private households

piped_water

number of permanent private households with water supply by piped network

well_spring_water

number of permanent private households with water supply by property's well or spring

stored_rainwater

number of permanent private households with water supply by stored rainwater

other_water_source

number of permanent private households with water supply by other source

private_bathroom

number of permanent private households with private bathroom or toilet

bathroom_sewerage

number of permanent private households with private bathroom or toilet & sanitation via sewerage or drainage network

bathroom_septic_tank

number of permanent private households with private bathroom or toilet & sanitation via septic tank

bathroom_cesspit

number of permanent private households with private bathroom or toilet & sanitation via cesspit

bathroom_ditch

number of permanent private households with private bathroom or toilet & wastewater discharged into ditch

bathroom_waterbodies

number of permanent private households with private bathroom or toilet & wastewater discharged into water bodies (river, lake or sea)

bathroom_other

number of permanent private households with private bathroom or toilet & wastewater discharged into other outlet